my-rag-service/requirements.txt
python-dotenv==1.*
langchain==0.*
langchain-core==0.*
langchain-community==0.*
langchain-openai==0.*
langgraph==0.*
pydantic==2.*
openai==1.*
tavily-python
fastapi
uvicorn
